implement close socket callbacks
